requirements.txt
streamlit
langchain
google-generativeai
chromadb
PyPDF2
python-dotenv